Enable shutdownOutput for EpollDomainSocketChannel
