+++ transport-native-epoll/src/main/java/io/netty/channel/unix/DomainSocketChannel.java
@@ -15,11 +15,13 @@
  */
 package io.netty.channel.unix;
 
+import io.netty.channel.socket.DuplexChannel;
+
 /**
  * A {@link UnixChannel} that supports communication via
  * <a href="http://en.wikipedia.org/wiki/Unix_domain_socket">Unix Domain Socket</a>.
  */
-public interface DomainSocketChannel extends UnixChannel {
+public interface DomainSocketChannel extends UnixChannel, DuplexChannel {
     @Override
     DomainSocketAddress remoteAddress();
